Don't bother checking pkg's REPOS_DIR
This check was originally added to ensure pkgbasify functioned properly
since we didn't use to override the user's REPOS_DIR config in
pkgbasify. However, the user's configured REPOS_DIR no longer affects
the operation of pkgbasify.

This check is fragile and broke recently due to a duplicate leading / in
the path checked against.

Sponsored by:	The FreeBSD Foundation
Closes: https://github.com/FreeBSDFoundation/pkgbasify/issues/38